Move transport constants to a new domains/sejours/phase1
