+++ packages/lib/src/constants/constants.ts
@@ -779,22 +779,7 @@ const TYPE_CLASSE = {
   FULL: "FULL",
 } as const;
 
-const TRANSPORT_MODES = {
-  TRAIN: "train",
-  BUS: "bus",
-  FUSEE: "fusée",
-  AVION: "avion",
-} as const;
 
-const TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT = 30;
-const TRANSPORT_CONVOCATION_SUBTRACT_MINUTES = {
-  [TRANSPORT_MODES.TRAIN]: 90,
-  [TRANSPORT_MODES.BUS]: TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
-  [TRANSPORT_MODES.FUSEE]: TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
-  [TRANSPORT_MODES.AVION]: TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
-} as const;
-
-const TRANSPORT_MODES_LIST = Object.values(TRANSPORT_MODES);
 
 const LIMIT_DATE_ESTIMATED_SEATS = new Date("2024-08-28");
 const LIMIT_DATE_TOTAL_SEATS = new Date("2024-09-30");
@@ -886,10 +871,6 @@ export {
   LIMIT_DATE_ESTIMATED_SEATS,
   LIMIT_DATE_TOTAL_SEATS,
   PHASE2_TOTAL_HOURS,
-  TRANSPORT_MODES_LIST,
-  TRANSPORT_MODES,
-  TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
-  TRANSPORT_CONVOCATION_SUBTRACT_MINUTES,
 };
 export default {
   YOUNG_STATUS,
@@ -956,8 +937,4 @@ export default {
   LIMIT_DATE_ESTIMATED_SEATS,
   LIMIT_DATE_TOTAL_SEATS,
   PHASE2_TOTAL_HOURS,
-  TRANSPORT_MODES_LIST,
-  TRANSPORT_MODES,
-  TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
-  TRANSPORT_CONVOCATION_SUBTRACT_MINUTES,
 };
