+++ packages/lib/src/domains/index.ts
@@ -1,0 +1,1 @@
+export * from './sejours/phase1/contants';
