+++ packages/lib/src/domains/sejours/phase1/contants.ts
@@ -1,0 +1,18 @@
+const TRANSPORT_MODES = {
+  TRAIN: "train",
+  BUS: "bus",
+  FUSEE: "fusée",
+  AVION: "avion",
+} as const;
+
+const TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT = 30;
+const TRANSPORT_CONVOCATION_SUBTRACT_MINUTES = {
+  [TRANSPORT_MODES.TRAIN]: 90,
+  [TRANSPORT_MODES.BUS]: TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
+  [TRANSPORT_MODES.FUSEE]: TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
+  [TRANSPORT_MODES.AVION]: TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT,
+} as const;
+
+const TRANSPORT_MODES_LIST = Object.values(TRANSPORT_MODES);
+
+export { TRANSPORT_MODES_LIST, TRANSPORT_MODES, TRANSPORT_CONVOCATION_SUBTRACT_MINUTES_DEFAULT, TRANSPORT_CONVOCATION_SUBTRACT_MINUTES };
