+++ packages/lib/src/index.ts
@@ -33,3 +33,4 @@ export * from "./utils/cohortUtils";
 export * from "./young";
 export * from "./mongoSchema";
 export * from "./constants/cohesionCenter";
+export * from "./domains";
